Fix macro recursion issues
The code previously only prevented replacement of the innermost macro being replaced, but mutual recursion of macros was not handled quite according to the standard (to my understanding, also checked with clang). This commit keeps a set of macros that have already been used in the current macro expansion, and refuses to expand them further.

The relevant part of the standard is:

> Then, the resulting preprocessing token sequence is rescanned, along with all subsequent preprocessing tokens of the source file, for more macro names to replace.
>
> If the name of the macro being replaced is found during this scan of the replacement list (not including the rest of the source file’s preprocessing tokens), it is not replaced. Furthermore, if any nested replacements encounter the name of the macro being replaced, it is not replaced. These nonreplaced macro name preprocessing tokens are no longer available for further replacement even if they are later (re)examined in contexts in which that macro name preprocessing token would otherwise have been replaced.